requirements.txt
flask==3.0.2
Werkzeug==3.0.2
hdfs==2.6.0
kafka-python
pyspark
pyarrow
